test(abg): move expected code for JAR to file
Thanks to this, it will also be tested if it's compilable. We can move
it because, unlikely to the test that checks the script-specific
generation, the file generated here is not meant to be used in a script.
